Create PROJ1_RUBRIC.md
Copy-pasted in the table we used for HW5, adjusted it to be generic; this is NOT our actual rubric assessment. Just wanted to create the file so we can all work on it. Will need to extend with the software sustainability stuff [here](https://github.com/Nikhil1912/Athenaeum/issues/4) as well. I.e. addressing issue #4